requirements.txt
torch>=1.4.0
torchvision>=0.5.0
dominate>=2.4.0
visdom>=0.1.8.8
wandb>=0.20.1

numpy>=1.24.4
pillow>=10.4.0
requests>=2.32.4
scipy>=1.10.1